+++ src/main/java/com/mcb/creditfactory/service/airplane/AirplaneAdapter.java
@@ -18,7 +18,7 @@ public class AirplaneAdapter implements CollateralObject {
     }
 
     @Override
-    public Short getYear() {
+    public Integer getYear() {
         return airplane.getYear();
     }
